Moved migration and docs
