+++ mkdocs.yml
@@ -146,7 +146,7 @@ nav:
     - Tools:
       - CLI: pages/tools/cli.md
       - Logging: pages/tools/logging.md
-      - Migration: pages/tools/migration.md
+    - Migration: pages/migration/index.md
     - Advanced Customizations:
       - Overview: pages/advanced-customization/index.md
       - Secrets Provider: pages/advanced-customization/secrets-provider-injection.md
